pr-summary-bot/pyproject.toml
[project]
name = "pr-summary-bot"
version = "0.1.0"
description = "PR summary bot powered by GitHub Copilot SDK"
requires-python = ">=3.12"
dependencies = [
    "github-copilot-sdk",
    "PyGithub>=2.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]
